requirements.txt
"urllib3<2"
standard-imghdr
python-telegram-bot==13.14
requests==2.32.5
python-dotenv==1.2.1
